Update balance pool logic

+++ spectrum-offchain-cardano/src/constants.rs
@@ -4,7 +4,7 @@ pub const FEE_DEN: u64 = 100000;
 
 pub const LEGACY_FEE_NUM_MULTIPLIER: u64 = 100;
 
-pub const WEIGHT_FEE_DEN: f64 = 5_f64;
+pub const WEIGHT_FEE_DEN: u64 = 5;
 
 pub const ADDITIONAL_ROUND_PRECISION: usize = 10;
